Move json formatting to cli subpackage
We're going to need this out of main in just a bit
